small refactor in modules to add TmdbServiceApiModule
