+++ app/src/main/java/com/arctouch/codechallenge/di/module/DetailsModule.kt
@@ -1,9 +1,9 @@
 package com.arctouch.codechallenge.di.module
 
-import com.arctouch.codechallenge.activity.details.DetailsActivity
-import com.arctouch.codechallenge.activity.details.DetailsPresenter
-import com.arctouch.codechallenge.activity.details.DetailsPresenterImpl
-import com.arctouch.codechallenge.activity.details.DetailsView
+import com.arctouch.codechallenge.ui.activity.details.DetailsActivity
+import com.arctouch.codechallenge.ui.activity.details.DetailsPresenter
+import com.arctouch.codechallenge.ui.activity.details.DetailsPresenterImpl
+import com.arctouch.codechallenge.ui.activity.details.DetailsView
 import com.jgabrielfreitas.infrastructure.api.TmdbApi
 import com.jgabrielfreitas.infrastructure.service.LocationConfigurationService
 import com.jgabrielfreitas.infrastructure.service.tmdb.BrazilianLocationConfigurationService
@@ -12,18 +12,8 @@ import com.jgabrielfreitas.infrastructure.service.tmdb.TmdbServiceImpl
 import dagger.Module
 import dagger.Provides
 
-@Module(includes = [(NetworkModule::class)])
+@Module(includes = [(NetworkModule::class), (TmdbServiceApiModule::class)])
 class DetailsModule {
-
-    @Provides
-    fun providesLocationConfiguration(): LocationConfigurationService {
-        return BrazilianLocationConfigurationService()
-    }
-
-    @Provides
-    fun providesTmdbService(serviceApi: TmdbApi, config: LocationConfigurationService): TmdbService {
-        return TmdbServiceImpl(serviceApi, config)
-    }
 
     @Provides
     fun providesView(detailsActivity: DetailsActivity): DetailsView {
