+++ app/src/main/java/com/arctouch/codechallenge/di/module/HomeModule.kt
@@ -1,9 +1,9 @@
 package com.arctouch.codechallenge.di.module
 
-import com.arctouch.codechallenge.activity.home.HomeActivity
-import com.arctouch.codechallenge.activity.home.HomePresenter
-import com.arctouch.codechallenge.activity.home.HomePresenterImpl
-import com.arctouch.codechallenge.activity.home.HomeView
+import com.arctouch.codechallenge.ui.activity.home.HomeActivity
+import com.arctouch.codechallenge.ui.activity.home.HomePresenter
+import com.arctouch.codechallenge.ui.activity.home.HomePresenterImpl
+import com.arctouch.codechallenge.ui.activity.home.HomeView
 import com.jgabrielfreitas.infrastructure.api.TmdbApi
 import com.jgabrielfreitas.infrastructure.service.LocationConfigurationService
 import com.jgabrielfreitas.infrastructure.service.tmdb.BrazilianLocationConfigurationService
@@ -12,18 +12,8 @@ import com.jgabrielfreitas.infrastructure.service.tmdb.TmdbServiceImpl
 import dagger.Module
 import dagger.Provides
 
-@Module(includes = [(NetworkModule::class)])
+@Module(includes = [(NetworkModule::class), (TmdbServiceApiModule::class)])
 class HomeModule {
-
-    @Provides
-    fun providesLocationConfiguration(): LocationConfigurationService {
-        return BrazilianLocationConfigurationService()
-    }
-
-    @Provides
-    fun providesTmdbService(serviceApi: TmdbApi, config: LocationConfigurationService): TmdbService {
-        return TmdbServiceImpl(serviceApi, config)
-    }
 
     @Provides
     fun providesView(homeActivity: HomeActivity): HomeView {
