+++ app/src/main/java/com/arctouch/codechallenge/di/module/TmdbServiceApiModule.kt
@@ -1,0 +1,24 @@
+package com.arctouch.codechallenge.di.module
+
+import com.jgabrielfreitas.infrastructure.api.TmdbApi
+import com.jgabrielfreitas.infrastructure.service.LocationConfigurationService
+import com.jgabrielfreitas.infrastructure.service.tmdb.BrazilianLocationConfigurationService
+import com.jgabrielfreitas.infrastructure.service.tmdb.TmdbService
+import com.jgabrielfreitas.infrastructure.service.tmdb.TmdbServiceImpl
+import dagger.Module
+import dagger.Provides
+
+@Module
+class TmdbServiceApiModule {
+
+    @Provides
+    fun providesLocationConfiguration(): LocationConfigurationService {
+        return BrazilianLocationConfigurationService()
+    }
+
+    @Provides
+    fun providesTmdbService(serviceApi: TmdbApi, config: LocationConfigurationService): TmdbService {
+        return TmdbServiceImpl(serviceApi, config)
+    }
+
+}
